refactor: entity 수정으로 인한 review 테이블 ì수정
