+++ src/main/java/com/ss/hanarowa/domain/lesson/service/ReviewService.java
@@ -3,5 +3,5 @@ package com.ss.hanarowa.domain.lesson.service;
 import com.ss.hanarowa.domain.lesson.dto.request.ReviewRequestDTO;
 
 public interface ReviewService {
-    void createReview(Long lessonId, Long memberId, ReviewRequestDTO reviewRequestDTO);
+    void createReview(Long lessonGisuId, Long memberId, ReviewRequestDTO reviewRequestDTO);
 }
