+++ src/main/java/com/ss/hanarowa/domain/member/repository/ReviewRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ss.hanarowa.domain.lesson.entity.LessonGisu;
-import com.ss.hanarowa.domain.member.entity.Review;
+import com.ss.hanarowa.domain.lesson.entity.Review;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
 	List<Review> findByLessonGisu(LessonGisu lessonGisu);
